:nail_care: refactor log time
