+++ tests/leave-status.spec.ts
@@ -6,7 +6,7 @@ import { identityURL, leaveURL } from "../src/utils/constants";
 import DashBoardPage from "../src/pages/leave/DashboardPage";
 import * as fs from "fs";
 
-test("Get Leave today list", async ({ page, request }) => {
+test("Get Leave today list", async ({ page }) => {
   await page.goto(identityURL);
 
   const loginPage = new LoginPage(page);
@@ -14,17 +14,8 @@ test("Get Leave today list", async ({ page, request }) => {
   const identityPage = new IdentityPage(page);
   await expect(identityPage.labelWelcome).toBeVisible();
   await page.goto(leaveURL);
-  const leavePage = new DashBoardPage(page, expect);
+  const leavePage = new DashBoardPage(page);
   const a = await leavePage.getLeaveTodayList();
   // save to file
   fs.writeFileSync("emp_off_today.txt", a.toString());
-  // send notification to slack
-  // await request.post(
-  //   "slackWebHookURL",
-  //   {
-  //     data: {
-  //       text: `Employees request off today (${a.length}):\n${a.toString()}`,
-  //     },
-  //   }
-  // );
 });
